make consistent with time.h
SCCS-vsn: 8.12

+++ sys/sys/stat.h
@@ -35,7 +35,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)stat.h	8.11 (Berkeley) 05/03/95
+ *	@(#)stat.h	8.12 (Berkeley) 06/16/95
  */
 
 #ifndef _SYS_STAT_H_
@@ -77,11 +77,11 @@ struct stat {
 	struct	timespec st_ctimespec;	/* time of last file status change */
 #else
 	time_t	  st_atime;		/* time of last access */
-	int32_t	  st_atimensec;		/* nsec of last access */
+	long	  st_atimensec;		/* nsec of last access */
 	time_t	  st_mtime;		/* time of last data modification */
-	int32_t	  st_mtimensec;		/* nsec of last data modification */
+	long	  st_mtimensec;		/* nsec of last data modification */
 	time_t	  st_ctime;		/* time of last file status change */
-	int32_t	  st_ctimensec;		/* nsec of last file status change */
+	long	  st_ctimensec;		/* nsec of last file status change */
 #endif
 	off_t	  st_size;		/* file size, in bytes */
 	int64_t	  st_blocks;		/* blocks allocated for file */
